refactor: rename arc-0200 assets feature
