+++ src/extension/apps/background/App.tsx
@@ -9,7 +9,7 @@ import Root from './Root';
 
 // features
 import { reducer as accountsReducer } from '@extension/features/accounts';
-import { reducer as arc200AssetsReducer } from '@extension/features/arc200-assets';
+import { reducer as arc200AssetsReducer } from '@extension/features/arc0200-assets';
 import { reducer as eventsReducer } from '@extension/features/events';
 import { reducer as messagesReducer } from '@extension/features/messages';
 import { reducer as networksReducer } from '@extension/features/networks';
@@ -28,7 +28,7 @@ const App: FC<IAppProps> = ({ i18next, initialColorMode }: IAppProps) => {
   const store: Store<IBackgroundRootState> = makeStore<IBackgroundRootState>(
     combineReducers({
       accounts: accountsReducer,
-      arc200Assets: arc200AssetsReducer,
+      arc0200Assets: arc200AssetsReducer,
       events: eventsReducer,
       messages: messagesReducer,
       networks: networksReducer,
