+++ src/extension/apps/registration/App.tsx
@@ -16,7 +16,7 @@ import {
 } from '@extension/constants';
 
 // features
-import { reducer as arc200AssetsReducer } from '@extension/features/arc200-assets';
+import { reducer as arc200AssetsReducer } from '@extension/features/arc0200-assets';
 import { reducer as networksReducer } from '@extension/features/networks';
 import { reducer as notificationsReducer } from '@extension/features/notifications';
 import { reducer as settingsReducer } from '@extension/features/settings';
@@ -66,7 +66,7 @@ const App: FC<IAppProps> = ({ i18next, initialColorMode }: IAppProps) => {
   const store: Store<IRegistrationRootState> =
     makeStore<IRegistrationRootState>(
       combineReducers({
-        arc200Assets: arc200AssetsReducer,
+        arc0200Assets: arc200AssetsReducer,
         networks: networksReducer,
         notifications: notificationsReducer,
         settings: settingsReducer,
