+++ src/extension/features/arc0200-assets/thunks/fetchARC0200AssetsFromStorageThunk.ts
