+++ src/extension/features/arc0200-assets/types/IARC0200AssetsState.ts
