+++ src/extension/features/arc0200-assets/types/IUpdateARC0200AssetInformationPayload.ts
