+++ src/extension/features/arc0200-assets/utils/getInitialState.ts
