+++ src/extension/selectors/arc-0200-assets/useSelectARC0200AssetsFetching.ts
@@ -5,6 +5,6 @@ import type { IMainRootState } from '@extension/types';
 
 export default function useSelectARC0200AssetsFetching(): boolean {
   return useSelector<IMainRootState, boolean>(
-    (state) => state.arc200Assets.fetching
+    (state) => state.arc0200Assets.fetching
   );
 }
