+++ src/extension/types/IBackgroundRootState.ts
@@ -1,18 +1,16 @@
 // features
-import { IAccountsState } from '@extension/features/accounts';
-import { IARC0200AssetsState } from '@extension/features/arc200-assets';
-import { IEventsState } from '@extension/features/events';
-import { INetworksState } from '@extension/features/networks';
-import { ISessionsState } from '@extension/features/sessions';
-import { ISettingsState } from '@extension/features/settings';
-import { IStandardAssetsState } from '@extension/features/standard-assets';
+import type { IAccountsState } from '@extension/features/accounts';
+import type { IEventsState } from '@extension/features/events';
+import type { INetworksState } from '@extension/features/networks';
+import type { ISessionsState } from '@extension/features/sessions';
+import type { ISettingsState } from '@extension/features/settings';
+import type { IStandardAssetsState } from '@extension/features/standard-assets';
 
 // types
-import IBaseRootState from './IBaseRootState';
+import type IBaseRootState from './IBaseRootState';
 
 interface IBackgroundRootState extends IBaseRootState {
   accounts: IAccountsState;
-  arc200Assets: IARC0200AssetsState;
   events: IEventsState;
   networks: INetworksState;
   sessions: ISessionsState;
