+++ src/extension/types/IBaseRootState.ts
@@ -1,7 +1,9 @@
 // features
-import { ISystemState } from '@extension/features/system';
+import type { IARC0200AssetsState } from '@extension/features/arc0200-assets';
+import type { ISystemState } from '@extension/features/system';
 
 interface IBaseRootState {
+  arc0200Assets: IARC0200AssetsState;
   system: ISystemState;
 }
