+++ tsconfig.json
@@ -35,8 +35,8 @@
       "@extension/features/arc0072-assets": [
         "src/extension/features/arc0072-assets"
       ],
-      "@extension/features/arc200-assets": [
-        "src/extension/features/arc200-assets"
+      "@extension/features/arc0200-assets": [
+        "src/extension/features/arc0200-assets"
       ],
       "@extension/features/events": ["src/extension/features/events"],
       "@extension/features/messages": ["src/extension/features/messages"],
